@@ -8,9 +8,11 @@ See https://quilljs.com/docs/modules/toolbar
 
 var React = require('react');
 var ReactDOMServer = require('react-dom/server');
+var createClass = require('create-react-class');
 var find = require('lodash/find');
 var isEqual = require('lodash/isEqual');
-var T = React.PropTypes;
+var T = require('prop-types');
+var DOM = require('react-dom-factories');
 
 var defaultColors = [
 	'rgb(  0,   0,   0)', 'rgb(230,   0,   0)', 'rgb(255, 153,   0)',
@@ -70,7 +72,7 @@ var defaultItems = [
 
 ];
 
-var QuillToolbar = React.createClass({
+var QuillToolbar = createClass({
 
 	displayName: 'Quill Toolbar',
 
@@ -183,7 +185,7 @@ var QuillToolbar = React.createClass({
 	render: function() {
 		var children = this.props.items.map(this.renderItem);
 		var html = children.map(ReactDOMServer.renderToStaticMarkup).join('');
-		return React.DOM.div({
+		return DOM.div({
 			id: this.props.id,
 			className: this.getClassName(),
 			style: this.props.style,
